Change wildcards parameter to file extension parameter
The reason for this change is that passing wildcards as parameters
easily caused problems when, for example, you used the additional
wildcard "*.md" and the current directory contained the file README.md,
then the shell would expand *.md to README.md and you'd be actually
passing -w README.md, which doesn't work.

Now we have a -e parameter which takes the file extension, e.g.
mow.py -e md, which can not be accidentally expanded by the shell